✅ Test : `@ParameterizedTest` for Counters
